Enable JIT on iOS.
In general, iOS does not support writing out unsigned executable code
in memory. This is enforced in two ways: first, pages cannot be both
writable and executable. Second, code must be signed by a trusted identity.

These restrictions are not in place for a process being debugged, as
debuggers need to execute code for interactive evaluation. However, only
processes with the "get-task-allow" entitlement can be debugged, and
that entitlement is only allowed with "development" provisioning profiles
that limit the number of devices they can be installed on.

Assuming Dolphin is signed with such an identity and entitlement, it is
possible to use enable the ARM JIT. This does require a few changes: to
enable writable + executable pages, `ptrace(PT_TRACE_ME, ...)` is
required to inform the kernel that the process is "being debugged".